Pass initial checksum value as uint16 to function

+++ util.go
@@ -1,8 +1,8 @@
 package motronic
 
 // Checksum16bit
-func Checksum16bit(initial uint32, start int, end int, buf []byte) uint16 {
-	var sum uint32 = initial
+func Checksum16bit(initial uint16, start int, end int, buf []byte) uint16 {
+	var sum uint32 = uint32(initial)
 
 	for i := start; i <= end; i += 1 {
 		sum += uint32(buf[i])
